created a gatekeeper file in our server folder, gave admin rights for make backend requests, fixed the login/signup bug for heroku
