+++ package.json
@@ -25,7 +25,7 @@
     "axios": "^0.21.1",
     "bcrypt": "^5.0.0",
     "compression": "^1.7.3",
-    "express": "^4.16.4",
+    "express": "^4.17.1",
     "history": "^4.9.0",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.9.1",
