+++ server/api/users.js
@@ -1,24 +1,7 @@
 const router = require('express').Router()
 const { models: { User }} = require('../db')
-module.exports = router
+const {requireToken, isAdmin} = require('./gatekeeper')
 
-const requireToken = async(req,res,next) => {
-  try {
-    const token = req.headers.authorization;
-    const user = await User.findByToken(token)
-    req.user = user
-    next()
-  } catch (error) {
-    next(error)
-  }
-}
-const isAdmin = (req,res,next) => {
-  if (req.user.userType !== 'admin'){
-    return res.status(403).send('You shall not pass!')
-  } else {
-    next ()
-  }
-}
 
 router.get('/', requireToken, isAdmin, async (req,res,next) => {
   try {
@@ -36,3 +19,5 @@ router.get('/', requireToken, isAdmin, async (req,res,next) => {
     next(error)
   }
 })
+
+module.exports = router
